requirements.txt
chatterbot
pyzmq
watchdog